+++ nemo_rl/models/generation/trtllm/trtllm_http_server.py
@@ -1,3 +1,16 @@
+# Copyright (c) 2026, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """OpenAI-compatible HTTP server wrapping ``tensorrt_llm.LLM``, serving /v1/chat/completions.
 
 Returns NeMoGym fields (prompt_token_ids, generation_token_ids, generation_log_probs).
@@ -19,10 +32,7 @@ from nemo_rl.models.generation.openai_server_utils import (
 
 logger = logging.getLogger(__name__)
 
-# Qwen3 tool-call delimiters — must match Qwen3ToolParser exactly so that
-# json.dumps normalisation and token boundaries are identical to the native server.
 _TOOL_BOT = "<tool_call>\n"
-_TOOL_EOT = "\n</tool_call>"
 _TOOL_CALL_RE = re.compile(
     r"<tool_call>\n(.*?)\n</tool_call>", re.DOTALL
 )
