+++ tests/unit/models/generation/test_vllm_generation.py
@@ -35,14 +35,12 @@ from nemo_rl.models.generation import configure_generation_config
 from nemo_rl.models.generation.interfaces import (
     GenerationDatumSpec,
 )
+from nemo_rl.models.generation.openai_server_utils import replace_prefix_tokens
 from nemo_rl.models.generation.vllm import VllmConfig, VllmGeneration
 from nemo_rl.models.generation.vllm.vllm_worker import (
     _resolve_enable_prefix_caching,
 )
-from nemo_rl.models.generation.vllm.vllm_worker_async import (
-    VllmAsyncGenerationWorkerImpl,
-    replace_prefix_tokens,
-)
+from nemo_rl.models.generation.vllm.vllm_worker_async import VllmAsyncGenerationWorkerImpl
 from nemo_rl.models.policy import LoRAConfig, PolicyConfig
 from nemo_rl.models.policy.lm_policy import Policy
 
@@ -1856,74 +1854,6 @@ def test_VllmAsyncGenerationWorker_replace_prefix_tokens(tokenizer):
         template_token_ids=template_token_ids,
     )
     assert result == model_token_ids
-
-
-def test_replace_prefix_tokens_empty_model_prefix_returns_template():
-    class _T:
-        eos_token_id = 2
-
-    tokenizer = _T()
-    model_prefix_token_ids = []
-    template_prefix_token_ids = [9, 2]
-    template_token_ids = [9, 2, 33, 44]
-    result = replace_prefix_tokens(
-        tokenizer=tokenizer,
-        model_prefix_token_ids=model_prefix_token_ids,
-        template_prefix_token_ids=template_prefix_token_ids,
-        template_token_ids=template_token_ids,
-    )
-    assert result == template_token_ids
-
-
-def test_replace_prefix_tokens_missing_eos_in_template_prefix_raises():
-    class _T:
-        eos_token_id = 2
-
-        def decode(self, *args, **kwargs):
-            pass
-
-    tokenizer = _T()
-    model_prefix_token_ids = [7, 2]
-    template_prefix_token_ids = [9, 9, 9]  # no EOS inside prefix
-    template_token_ids = [9, 9, 9, 2, 10]
-    with pytest.raises(AssertionError):
-        replace_prefix_tokens(
-            tokenizer=tokenizer,
-            model_prefix_token_ids=model_prefix_token_ids,
-            template_prefix_token_ids=template_prefix_token_ids,
-            template_token_ids=template_token_ids,
-        )
-
-
-def test_replace_prefix_tokens_tokenizer_without_eos_raises():
-    class _T:
-        eos_token_id = None
-
-    tokenizer = _T()
-    with pytest.raises(AssertionError):
-        replace_prefix_tokens(
-            tokenizer=tokenizer,
-            model_prefix_token_ids=[1],
-            template_prefix_token_ids=[1, 2],
-            template_token_ids=[1, 2],
-        )
-
-
-def test_replace_prefix_tokens_uses_last_eos_in_template_prefix():
-    class _T:
-        eos_token_id = 2
-
-    tokenizer = _T()
-    model_prefix_token_ids = [100, 2]
-    template_prefix_token_ids = [9, 2, 9, 2]  # two EOS; last at idx=3
-    template_token_ids = [9, 2, 9, 2, 77, 88]
-    result = replace_prefix_tokens(
-        tokenizer=tokenizer,
-        model_prefix_token_ids=model_prefix_token_ids,
-        template_prefix_token_ids=template_prefix_token_ids,
-        template_token_ids=template_token_ids,
-    )
-    assert result == [100, 2, 77, 88]
 
 
 @pytest.mark.asyncio
